pyproject.toml
[tool.poetry]
name = "WEB_HT12_REST_API_AUTH"
version = "0.1.0"
description = "REST API for contacts saving. Only authorized users can add new contacts."
authors = ["Yuliia Melnychenko <arwen.vogel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.20.0"
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy = "^1.4.46"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"
python-multipart = "^0.0.5"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
psycopg2-binary  = "^2.9.9"
pydantic = {version = "<2.0.0", extras = ["email"]} 


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
